clickhouse-chatbot/requirements.txt
clickhouse-driver
websockets
pyyaml

